requirements.txt
numpy==1.26.2
pandas==2.1.4
PyMySQL==1.1.0
SQLAlchemy==2.0.23
PyYAML==6.0.1